Move play page to multiplayer feature directory
